Move trust args to subcommands + move cli args to commands/mod.rs
